moved ApplicationOptions
