+++ TagsCloudVisualization/Program.cs
@@ -1,6 +1,7 @@
 ﻿using System.Drawing.Imaging;
 using Autofac;
 using CommandLine;
+using TagsCloudVisualization.ApplicationOptions;
 using TagsCloudVisualization.Visualization;
 
 namespace TagsCloudVisualization
@@ -9,7 +10,7 @@ namespace TagsCloudVisualization
     {
         public static void Main(string[] args)
         {
-            var parsedArguments = Parser.Default.ParseArguments<ApplicationOptions>(args);
+            var parsedArguments = Parser.Default.ParseArguments<ApplicationOptions.ApplicationOptions>(args);
             var applicationOptions = new ApplicationOptionsExtractor().GetOption(parsedArguments);
             var container = new ContainerCreator().GetContainer(applicationOptions.GetVisualizingOptions());
             var cloudCreator = container.Resolve<CloudCreator>();
